Rename Ruby requirement class
